Move email change handling from helper functions to a class.
This follows the pattern of the PendingUser interactions with the
sqlite database, and should be easier to extend and otherwise work
with.

Introduce a new intermediate class which handles the age stuff and
then a PendingEmail class which extends that. Also unit tests for
this class. Remove the previous helper methods for email changing,
and replace them with instances of this class.

Change-Id: I579ddfade0e9ef80cb2c129672150e14a784ce44