Fix #2183: Reject emails that can't work.
This modifies the registration page (only) to reject emails that don't
conform to the general pattern of a@b.cc. This does not ensure that
the email is valid, but prevents wildly invalid data being entered.
Validation is done both in JS and on the backend, and the input field
is updated to make use of HTML5 validation too.

Change-Id: I8263dc9b8ecfedc6b10939f6234585e1a4df831e