Cut back the information in the initial competitor email.
This reduces the first email that competitors receive to being a
minimal instruction that they need to activate their account. While
it still tells them why they'd want an account, and that there's
one waiting for them, it no longer tells them anything about the
accout itself.

Their email and username are no longer listed in the email (though
the username remains part of the activation url; most people probably
won't notice that).

The information that the link is only valid for a short time is
also now more prominent.

Change-Id: Ie4e775b7ebbba5c7e9f31cf2b407fb67272567ac