Support validation of recipients during preview
This changes the form of the recipients field on the email template
to a list (based on comma separated input) and then runs a validator
over those during the preview. Any unrecognised entries are shown
as errors.

Formatting errors are no longer fatal to the production of the
preview and will be shown in the same way as invalid recipients.

Change-Id: I4b6001d1a2652e00aabd3d5bfe87c9c31b2c85fe